+++ changelog.md
@@ -1,3 +1,10 @@
+# Version 6.0.0 - dd.mm.2024
+
+## Features
+
+- added ``habapp_rules.sensors.humidity.HumiditySwitch`` to set a switch item if high humidity is detected. Currently only a absolut threshold is accepted
+- send update of summer / winter of ``habapp_rules.system.summer_winter.SummerWinter`` after every check. If this rule is used to send the summer / winter state to the KNX bus, this ensures, that the state is sent at least once a day
+
 # Version 5.7.0 - 09.04.2024
 
 ## Features